Add files via upload
This is the original setup I have for converting PDF documents to TXT documents by reading the PDF document, then writing to a TXT document. I have plans to build on this program by further filtering the data from the PDF documents and building a database with some of the critical data that will be read.